+++ server/src/api/controllers/sessionsController.ts
@@ -6,8 +6,10 @@ import {
   Path,
   Post,
   Put,
+  Request,
   Response,
   Route,
+  Security,
 } from 'tsoa';
 
 import {
@@ -17,7 +19,12 @@ import {
 } from '../../services/session';
 
 import type { Session } from '../../../generated/prisma';
-import type { ValidateErrorJSON } from '../../types/errors';
+import type {
+  ForbiddenErrorJson,
+  UnauthorizedErrorJSON,
+  ValidateErrorJSON,
+} from '../../types/errors';
+import type { Request as ExRequest } from 'express';
 
 @Route('sessions')
 export class SessionsController extends Controller {
@@ -26,11 +33,25 @@ export class SessionsController extends Controller {
     return await new SessionService().getById(sessionId);
   }
 
+  @Security('session_auth')
   @Response<ValidateErrorJSON>(422, 'Validation Failed')
+  @Response<UnauthorizedErrorJSON>(401, 'Unauthorized')
+  @Response<ForbiddenErrorJson>(403, 'Forbidden')
   @Post()
   public async createSession(
     @Body() requestBody: SessionCreationParams,
+    @Request() exReq: ExRequest,
   ): Promise<Session> {
+    if (!exReq.session.participantId) {
+      this.setStatus(401);
+      throw new Error('Unauthorized');
+    }
+
+    if (!exReq.session.isAdmin) {
+      this.setStatus(403);
+      throw new Error('Only admin can create session');
+    }
+
     return await new SessionService().create({
       title: requestBody.title,
       code: requestBody.code,
@@ -39,17 +60,47 @@ export class SessionsController extends Controller {
     });
   }
 
+  @Security('session_auth')
   @Response<ValidateErrorJSON>(422, 'Validation Failed')
+  @Response<UnauthorizedErrorJSON>(401, 'Unauthorized')
+  @Response<ForbiddenErrorJson>(403, 'Forbidden')
   @Put('{sessionId}')
   public async updateSession(
     @Path() sessionId: string,
     @Body() requestBody: SessionUpdateParams,
+    @Request() exReq: ExRequest,
   ): Promise<Session> {
+    if (!exReq.session.participantId) {
+      this.setStatus(401);
+      throw new Error('Unauthorized');
+    }
+
+    if (!exReq.session.isAdmin) {
+      this.setStatus(403);
+      throw new Error('Only admin can create session');
+    }
+
     return await new SessionService().update(sessionId, requestBody);
   }
 
+  @Security('session_auth')
+  @Response<UnauthorizedErrorJSON>(401, 'Unauthorized')
+  @Response<ForbiddenErrorJson>(403, 'Forbidden')
   @Delete('{sessionId}')
-  public async deleteSession(@Path() sessionId: string): Promise<Session> {
+  public async deleteSession(
+    @Path() sessionId: string,
+    @Request() exReq: ExRequest,
+  ): Promise<Session> {
+    if (!exReq.session.participantId) {
+      this.setStatus(401);
+      throw new Error('Unauthorized');
+    }
+
+    if (!exReq.session.isAdmin) {
+      this.setStatus(403);
+      throw new Error('Only admin can delete session');
+    }
+
     return await new SessionService().delete(sessionId);
   }
 }
